+++ packages/cloud-cognitive/src/components/Datagrid/useExpandedRow.js
@@ -22,7 +22,6 @@ const useExpandedRow = (hooks) => {
       canExpand: row.original && !row.original.notExpandable,
       expandedContentHeight:
         expandedRowsHeight[row.index] || expandedContentHeight,
-      // RowRenderer: DatagridExpandedRow(row.RowRenderer, expansionRenderer),
       RowRenderer: DatagridExpandedRow(
         row.RowRenderer,
         ExpandedRowContentComponent