+++ packages/cloud-cognitive/src/components/Datagrid/utils/makeData.js
@@ -7,6 +7,21 @@
 
 import React from 'react';
 import namor from 'namor';
+import { StatusIcon } from '../../StatusIcon';
+import { inlineEditSelectItems } from './getInlineEditColumns';
+
+const getRandomInteger = (min, max, decimalPlaces) => {
+  const roundedMin = Math.ceil(min);
+  const roundedMax = Math.floor(max);
+  const randomNumber = Math.random() * (max - min) + min;
+  if (!decimalPlaces) {
+    return (
+      Math.floor(Math.random() * (roundedMax - roundedMin + 1)) + roundedMin
+    );
+  }
+  const power = Math.pow(10, decimalPlaces);
+  return Math.floor(randomNumber * power) / power;
+};
 
 export const makeData = (...lens) => {
   const makeDataLevel = (depth = 0) => {
@@ -28,8 +43,34 @@ export const range = (len) => {
   return arr;
 };
 
+const renderStatusIcon = (statusChance) => {
+  const iconProps = {
+    size: 'sm',
+    theme: 'light',
+    kind:
+      statusChance > 0.66
+        ? 'critical'
+        : statusChance > 0.33
+        ? 'minor-warning'
+        : 'normal',
+    iconDescription:
+      statusChance > 0.66
+        ? 'Critical'
+        : statusChance > 0.33
+        ? 'Minor warning'
+        : 'Normal',
+  };
+  return <StatusIcon {...iconProps} />;
+};
+
 const newPerson = () => {
   const statusChance = Math.random();
+  const initialChartTypeIndex = getRandomInteger(0, 2);
+  const activeSinceDate = new Date();
+  let yesterdayDate = new Date();
+  yesterdayDate.setDate(yesterdayDate.getDate() - 1);
+  let twoDaysAgoDate = new Date();
+  twoDaysAgoDate.setDate(twoDaysAgoDate.getDate() - 2);
   return {
     firstName: namor.generate({ words: 1, numbers: 0 }),
     lastName: namor.generate({ words: 1, numbers: 0 }),
@@ -62,6 +103,20 @@ const newPerson = () => {
         : statusChance > 0.33
         ? 'complicated'
         : 'single',
+    chartType:
+      initialChartTypeIndex === 0
+        ? inlineEditSelectItems[0]
+        : initialChartTypeIndex === 1
+        ? inlineEditSelectItems[1]
+        : inlineEditSelectItems[2],
+    activeSince:
+      statusChance > 0.66
+        ? activeSinceDate
+        : statusChance > 0.33
+        ? yesterdayDate
+        : twoDaysAgoDate,
+    bonus: `$\r${getRandomInteger(100, 500, 2)}`,
+    status_icon: renderStatusIcon(statusChance),
   };
 };